Splitting test file into QGL and non-QGL test files
