+++ tests/geometry/tools/testSphericalAccumulator.cpp
@@ -41,87 +41,6 @@ using namespace DGtal;
 // Functions for testing class SphericalAccumulator.
 ///////////////////////////////////////////////////////////////////////////////
 
-
-
-#ifdef WITH_VISU3D_QGLVIEWER
-#include <QtGui/qapplication.h>
-#include "DGtal/io/viewers/Viewer3D.h"
-bool testSphericalViewer(int argc, char **argv)
-{
-  QApplication application(argc,argv);
-  unsigned int nbok = 0;
-  unsigned int nb = 0;
-  
-  trace.beginBlock ( "Testing Spherical Accumulator Viewer..." );
-  
-  typedef Z3i::RealVector Vector;
-  
-  SphericalAccumulator<Vector> accumulator(15);
-  trace.info()<< accumulator << std::endl;
-  
-  for(unsigned int i=0; i< 10000; i++)
-    accumulator.addDirection( Vector (1+10.0*(rand()-RAND_MAX/2)/(double)RAND_MAX,
-				      (1+10.0*(rand()-RAND_MAX/2))/(double)RAND_MAX,
-				      (1+10.0*(rand()-RAND_MAX/2))/(double)RAND_MAX));
-  
-  Viewer3D viewer;
-  viewer.show();
-  Vector a,b,c,d;
-  viewer << accumulator;
-
-    trace.info() << "Bin values: ";
-  for(SphericalAccumulator<Vector>::ConstIterator it=accumulator.begin(), itend=accumulator.end();
-      it != itend;
-      ++it)
-    trace.info() << *it<<" ";
-  trace.info() << std::endl;
-  trace.info() << accumulator<<std::endl;
-
-  viewer << Viewer3D::updateDisplay;
-  bool res = application.exec();
-  trace.emphase() << ( res ? "Passed." : "Error." ) << endl;
-  trace.endBlock();
-  return res ? 0 : 1;
-}
-bool testSphericalViewerInteger(int argc, char **argv)
-{
-  QApplication application(argc,argv);
-  unsigned int nbok = 0;
-  unsigned int nb = 0;
-  
-  trace.beginBlock ( "Testing Spherical Accumulator Viewer  with Integer numbers..." );
-  
-  typedef Z3i::Vector Vector;
-  
-  SphericalAccumulator<Vector> accumulator(15);
-  trace.info()<< accumulator << std::endl;
-  
-  for(unsigned int i=0; i< 10000; i++)
-    accumulator.addDirection( Vector (1+(rand()-RAND_MAX/2),
-                                      (1+(rand()-RAND_MAX/2)),
-                                      (1+(rand()-RAND_MAX/2))));
-  
-  Viewer3D viewer;
-  viewer.show();
-  Vector a,b,c,d;
-  Display3DFactory::draw(viewer,accumulator, Z3i::RealVector(1.0,1.0,1.0), 3.0);
-
-    trace.info() << "Bin values: ";
-  for(SphericalAccumulator<Vector>::ConstIterator it=accumulator.begin(), itend=accumulator.end();
-      it != itend;
-      ++it)
-    trace.info() << *it<<" ";
-  trace.info() << std::endl;
-  trace.info() << accumulator<<std::endl;
-
-  viewer << Viewer3D::updateDisplay;
-  bool res = application.exec();
-  trace.emphase() << ( res ? "Passed." : "Error." ) << endl;
-  trace.endBlock();
-  return res ? 0 : 1;
-}
-#endif
-
 bool testSphericalAccumulator()
 {
   unsigned int nbok = 0;
@@ -293,10 +212,6 @@ int main( int argc, char** argv )
   trace.info() << endl;
 
   bool res = testSphericalAccumulator() && testSphericalMore()
-#ifdef WITH_VISU3D_QGLVIEWER
-    && testSphericalViewer(argc,argv)
-    && testSphericalViewerInteger(argc,argv)
-#endif 
     && testSphericalMoreIntegerDir();
   trace.emphase() << ( res ? "Passed." : "Error." ) << endl;
   trace.endBlock();
